chore: move bundler file to runtimes/node
