+++ src/runtimes/node/index.js
@@ -4,11 +4,14 @@ const cpFile = require('cp-file')
 
 const { JS_BUNDLER_ESBUILD, JS_BUNDLER_ESBUILD_ZISI, JS_BUNDLER_ZISI, RUNTIME_JS } = require('../../utils/consts')
 
-const { getDefaultBundler } = require('./default_bundler')
 const { findFunctionsInPaths } = require('./finder')
 const { getSrcFiles } = require('./src_files')
 const { zipEsbuild } = require('./zip_esbuild')
 const { zipZisi } = require('./zip_zisi')
+
+// We use ZISI as the default bundler until the next major release, with the
+// exception of TypeScript files, for which the only option is esbuild.
+const getDefaultBundler = ({ extension }) => (extension === '.ts' ? JS_BUNDLER_ESBUILD : JS_BUNDLER_ZISI)
 
 const zipFunction = async function ({
   destFolder,
